+++ Engine/src/game_objects/animation/Bone.cpp
@@ -10,10 +10,10 @@ Bone::Bone(const std::string& name, int ID, const aiNodeAnim* channel) :
 	for (int positionIndex = 0; positionIndex < m_NumPositions; ++positionIndex)
 	{
 		aiVector3D aiPosition = channel->mPositionKeys[positionIndex].mValue;
-		float timeStamp = channel->mPositionKeys[positionIndex].mTime;
+		double timeStamp = channel->mPositionKeys[positionIndex].mTime;
 		KeyPosition data;
 		data.position = glm::vec3(aiPosition.x, aiPosition.y, aiPosition.z);
-		data.timeStamp = timeStamp;
+		data.timeStamp = (float)timeStamp;
 		m_Positions.push_back(data);
 	}
 
@@ -21,7 +21,7 @@ Bone::Bone(const std::string& name, int ID, const aiNodeAnim* channel) :
 	for (int rotationIndex = 0; rotationIndex < m_NumRotations; ++rotationIndex)
 	{
 		aiQuaternion aiOrientation = channel->mRotationKeys[rotationIndex].mValue;
-		float timeStamp = channel->mRotationKeys[rotationIndex].mTime;
+		double timeStamp = channel->mRotationKeys[rotationIndex].mTime;
 		KeyRotation data;
 		data.orientation = glm::vec3(aiOrientation.x, aiOrientation.y, aiOrientation.z);
 		data.timeStamp = timeStamp;
@@ -32,7 +32,7 @@ Bone::Bone(const std::string& name, int ID, const aiNodeAnim* channel) :
 	for (int keyIndex = 0; keyIndex < m_NumScalings; ++keyIndex)
 	{
 		aiVector3D scale = channel->mScalingKeys[keyIndex].mValue;
-		float timeStamp = channel->mScalingKeys[keyIndex].mTime;
+		double timeStamp = channel->mScalingKeys[keyIndex].mTime;
 		KeyScale data;
 		data.scale = glm::vec3(scale.x, scale.y, scale.z);
 		data.timeStamp = timeStamp;
